Move test to extension module to increase coverage
